+++ lib/src/websocket_request_manager/interfaces/socket_object.dart
@@ -1,4 +1,0 @@
-abstract class ISocketObject<T> {
-  DateTime get timestamp;
-  T get object;
-}