+++ lib/src/websocket_request_manager/interfaces/socket_request.dart
@@ -7,8 +7,13 @@ abstract class ISocketRequest {
   ISocketMessage get requestMessage;
 
   /// Timeout time in milliseconds. If this time has passed and all required
-  ///
+  /// [responseTopics] are not received - request fails.
   int get timeoutMs;
+
+  /// If [cacheTimeMs] hasn't passed since data was received then get stored
+  /// value, otherwise - request and wait for new value. On null data is never
+  /// obsolete.
+  int? get cacheTimeMs;
 
   /// Multiple answers. When all answer topics are received,
   /// then the request is considered successful.