+++ lib/src/websocket_request_manager/interfaces/websocket_request_manager.dart
@@ -4,12 +4,24 @@ import '../../../websocket_universal.dart';
 
 /// asyncSocket.requestData(`socketTopic`)
 abstract class IWebSocketRequestManager {
-  Stream<ITimedMessage> get decodedMessagesStream;
-  Stream<IFinishedSocketRequest> get finishedRequestsStream;
+  /// Single messages
+  Stream<ITimedSocketResponse> get decodedMessagesStream;
+
+  /// Composite responses stream
+  Stream<ICompositeSocketResponse> get finishedRequestsStream;
+
+  /// Ping stream
   Stream<int> get pingMsStream;
 
-  void requestData(ISocketRequest socketRequest);
-  ITimedMessage? getStoredDecodedMessage(String key);
+  /// Last ping delay
+  int get pingDelayMs;
 
+  /// Send request to server
+  void requestData(ISocketRequest socketRequest);
+
+  /// Get cached server response
+  ITimedSocketResponse? getStoredDecodedMessage(String key);
+
+  /// Dispose
   void close();
 }